+++ nft-storage/src/lib.rs
@@ -4,6 +4,9 @@ use reqwest::{ Client, Body };
 use chrono::{Duration, Utc, SecondsFormat};
 
 mod utils;
+mod cors;
+
+type CorsHeaders = cors::CorsHeaders;
 
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -39,28 +42,7 @@ struct UrlPinRequest {
 }
 
 
-#[derive(Debug)]
-struct CorsHeaders {}
 
-impl CorsHeaders {
-    pub fn new() -> Result<Headers> {
-        let mut headers = Headers::new();
-        headers.set("Access-Control-Allow-Origin", "*")?;
-        headers.set("Access-Control-Allow-Methods", "GET, HEAD, POST, OPTIONS")?;
-        headers.set("Access-Control-Allow-Headers", "*")?;
-        return Ok(headers);
-    }
-
-    pub fn response() -> Result<Response> {
-        let resp = Response::empty()?.with_headers(CorsHeaders::new()?);
-        return Ok(resp);
-    }
-
-    pub fn update(response: Result<Response>) -> Result<Response> {
-        let resp = response?.with_headers(CorsHeaders::new()?);
-        return Ok(resp);
-    }
-}
 
 fn log_request(req: &Request) {
     console_log!(
